libknc: major API rototill.
We are starting to make a major rototill to the libknc API to bring it
into a state where we can commit to it.  This commit introduces new
constructors, removes old constructors, allows one to set and query
various properties of the knc_ctx, renames struct knc_ctx * into knc_ctx,
and much else.  It is not quite complete insofar as the API is still
not exactly that to which we would prefer to commit, however, it is
much closer and the [modified] unit test still works.