libknc: a bit of help on TCP errors.
We modify knc_{fill,flush} to return errno = EIO if there
is a pending error in the KNC context.

We also mark an EOF read as an error condition as the KNC
session has not been properly terminated with a command
(not implemented, yet...)

We continue if we are in blocking mode and get an EAGAIN---this
should likely be some sort of error as we don't want to busy loop
here, but we'll deal with that later.

And finally, we return errno=EPIPE from knc_write if the socket is
closed.