libknc: allow multiple commands in a single packet.
We now process multiple commands in a single packet but we still only send
one.  Not because of compatibility as this is still an unpublished interface
but rather for simplicity of implementation at the current time.  We will
expand this later and it will likely involve some level of coalescing.